+++ includes/helpers/image-functions.php
@@ -19,3 +19,22 @@ function get_all_image_sizes()
 	/* Merge all Image Sizes */
 	return array_merge($image_sizes, $additional_image_sizes);
 }
+
+/**
+ * Get image height
+ */
+function ejo_get_image_dimensions( $image_id, $image_size = 'full' )
+{
+	$image = wp_get_attachment_image_src( $image_id , $image_size );
+
+	if ( !$image )
+		return false;
+
+	$image_dimensions = array();
+	$image_dimensions['width'] = $image[1];
+	$image_dimensions['height'] = $image[2];
+	$image_dimensions['src'] = $image[0];
+	$image_dimensions['id'] = $image_id;
+
+	return $image_dimensions;
+}